backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
dj-database-url==2.3.0
psycopg[binary]==3.2.9
python-dotenv==1.1.0
requests==2.32.3
